wasteless-ai/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlmodel==0.0.21
openai==1.53.0
httpx==0.27.2